Fix build break when doing 32-bit build with some certain C compiler option combinations.
Use the library functions for shift operations in BaseLib for a 64-bit integer where the code is shared for 32-bit and 64-bit.
Defining bitfields in structures with > 32 bits will cause these types of issues on IA32 builds. So the largest bitfield should be type UINT32 with a max size of :32.

git-svn-id: svn://svn.code.sf.net/p/edk2/code/trunk/edk2/UefiCpuPkg@10983 6f19259b-4bc3-4df7-8a09-765794883524